Rip out http solver and legacy adapter
