+++ crates/e2e/src/setup/colocation.rs
@@ -1,6 +1,6 @@
 use {
     crate::{nodes::NODE_HOST, setup::*},
-    ethcontract::{H160, U256},
+    ethcontract::H160,
     reqwest::Url,
     tokio::task::JoinHandle,
 };
@@ -17,19 +17,6 @@ native-token-price-estimation-amount = "100000000000000000"
     ));
 
     start_solver(config_file, "baseline".to_string()).await
-}
-
-pub async fn start_legacy_solver(solver_endpoint: Url, chain_id: Option<U256>) -> Url {
-    let chain_id = chain_id.unwrap_or(U256::from(1));
-    let config_file = config_tmp_file(format!(
-        r#"
-chain-id = "{chain_id}"
-solver-name = "legacy"
-endpoint = "{solver_endpoint}"
-        "#,
-    ));
-
-    start_solver(config_file, "legacy".to_string()).await
 }
 
 async fn start_solver(config_file: TempPath, solver_name: String) -> Url {
