+++ crates/solvers/src/api/mod.rs
@@ -28,7 +28,6 @@ impl Api {
             .route("/metrics", axum::routing::get(routes::metrics))
             .route("/healthz", axum::routing::get(routes::healthz))
             .route("/solve", axum::routing::post(routes::solve))
-            .route("/notify", axum::routing::post(routes::notify))
             .layer(
                 tower::ServiceBuilder::new().layer(tower_http::trace::TraceLayer::new_for_http()),
             )
