+++ crates/solvers/src/api/routes/mod.rs
@@ -2,10 +2,9 @@ use serde::Serialize;
 
 mod healthz;
 mod metrics;
-mod notify;
 mod solve;
 
-pub(super) use {healthz::healthz, metrics::metrics, notify::notify, solve::solve};
+pub(super) use {healthz::healthz, metrics::metrics, solve::solve};
 
 #[derive(Debug, Serialize)]
 #[serde(untagged)]
