+++ crates/solvers/src/boundary/mod.rs
@@ -2,7 +2,6 @@
 //! "legacy" logic in `shared` and `model`.
 
 pub mod baseline;
-pub mod legacy;
 pub mod liquidity;
 pub mod naive;
