+++ crates/solvers/src/domain/solver/mod.rs
@@ -1,18 +1,16 @@
 use crate::{
-    domain::{auction, notification, solution},
+    domain::{auction, solution},
     infra::metrics,
 };
 
 pub mod baseline;
-pub mod legacy;
 pub mod naive;
 
-pub use self::{baseline::Baseline, legacy::Legacy, naive::Naive};
+pub use self::{baseline::Baseline, naive::Naive};
 
 pub enum Solver {
     Baseline(Baseline),
     Naive(Naive),
-    Legacy(Legacy),
 }
 
 impl Solver {
@@ -25,19 +23,8 @@ impl Solver {
         let solutions = match self {
             Solver::Baseline(solver) => solver.solve(auction).await,
             Solver::Naive(solver) => solver.solve(auction).await,
-            Solver::Legacy(solver) => solver.solve(auction).await,
         };
         metrics::solved(&deadline, &solutions);
         solutions
-    }
-
-    /// Notifies the solver about important events. Some of those events are
-    /// generated by the solver engine, others by the driver.
-    pub fn notify(&self, notification: notification::Notification) {
-        match self {
-            Solver::Baseline(_) => (),
-            Solver::Naive(_) => (),
-            Solver::Legacy(solver) => solver.notify(notification),
-        }
     }
 }
