+++ crates/solvers/src/infra/config/mod.rs
@@ -1,7 +1,6 @@
 use std::fmt::Debug;
 
 pub mod baseline;
-pub mod legacy;
 
 /// Unwraps result or logs a `TOML` parsing error.
 fn unwrap_or_log<T, E, P>(result: Result<T, E>, path: &P) -> T
