+++ crates/solvers/src/infra/mod.rs
@@ -3,4 +3,3 @@ pub mod cli;
 pub mod config;
 pub mod contracts;
 pub mod metrics;
-pub mod persistence;
